requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.1.0
SpeechRecognition==3.10.0
gTTS==2.5.0
googletrans==4.0.0rc1
geopy==2.4.1